# docs/swagger.yaml
basePath: /
definitions:
  schema.CreateExchangeInput:
    properties:
      event_id:
        type: string
      user_id_1:
        type: string
      user_id_2:
        type: string
    type: object
  schema.Exchange:
    properties:
      event_id:
        type: string
      user_id:
        type: string
    type: object
  schema.GetExchangesOutput:
    properties:
      exchanges:
        items:
          $ref: '#/definitions/schema.Exchange'
        type: array
    type: object
  schema.GetMeOutput:
    properties:
      image_url:
        type: string
      is_organizer:
        type: boolean
      message:
        type: string
      name:
        type: string
      skills:
        additionalProperties:
          type: string
        type: object
      urls:
        additionalProperties:
          type: string
        type: object
      user_id:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /exchanges:
    get:
      consumes:
      - application/json
      description: 名刺交換の一覧を取得するためのエンドポイント
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.GetExchangesOutput'
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: 交換した名刺一覧取得
      tags:
      - Exchange
    post:
      consumes:
      - application/json
      description: 名刺交換のためのエンドポイント
      parameters:
      - description: Exchange request body
        in: body
        name: exchange
        required: true
        schema:
          $ref: '#/definitions/schema.CreateExchangeInput'
      produces:
      - application/json
      responses:
        "201":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: 名刺交換
      tags:
      - Exchange
  /me:
    get:
      consumes:
      - application/json
      description: 自分の情報を取得する
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.GetMeOutput'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: 自分の情報
      tags:
      - User
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"